pyproject.toml
[project]
name = "lol-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.52.1",
    "pillow>=10.0.0",
    "langchain>=1.1.3",
    "langchain-core>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=0.1.0",
    "ddgs>=9.9.3",
    "loguru>=0.7.3",
    "trafilatura>=2.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.1.0",
    "torch>=2.0.0",
    "python-dotenv>=1.0.0",
]
